feat: copy public html test structor
