+++ tests/integration/clean-dist-path/tests/index.test.ts
@@ -3,7 +3,7 @@ import { fs } from '@modern-js/utils';
 import { modernBuild } from '../../../utils/modernTestUtils';
 
 describe('clean dist path', () => {
-  it(`should not clean dist path when output.cleanDistPath is false`, async () => {
+  test(`should not clean dist path when output.cleanDistPath is false`, async () => {
     const appDir = path.resolve(__dirname, '..');
     const tempFile = path.join(appDir, 'dist/foo.txt');
     const htmlFile = path.join(appDir, 'dist/html/main/index.html');
