+++ tests/integration/copy-assets/config/public/demo.html
