--- tests/integration/copy-assets/fixtures/copy-public-html/package.json
@@ -1,16 +1,0 @@
-{
-  "private": true,
-  "name": "integration-copy-public-html",
-  "version": "2.9.0",
-  "scripts": {
-    "build": "modern build"
-  },
-  "dependencies": {
-    "react": "^18",
-    "react-dom": "^18",
-    "@babel/runtime": "^7.18.0"
-  },
-  "devDependencies": {
-    "@modern-js/app-tools": "workspace:*"
-  }
-}
