+++ tests/integration/copy-assets/modern.config.ts
