+++ tests/integration/copy-assets/package.json
@@ -3,6 +3,12 @@
   "name": "integration-copy-assets",
   "version": "2.9.0",
   "dependencies": {
+    "react": "^18",
+    "react-dom": "^18",
+    "@babel/runtime": "^7.18.0"
+  },
+  "devDependencies": {
+    "@modern-js/app-tools": "workspace:*",
     "@types/jest": "^29",
     "@types/node": "^14"
   }
