+++ tests/integration/copy-assets/src/index.ts
