+++ tests/integration/copy-assets/tests/.eslintrc.js
