+++ tests/integration/copy-assets/tests/rspack.test.ts
@@ -1,0 +1,7 @@
+import { testPublicHtml } from './testPublicHtml';
+
+describe('copy assets', () => {
+  test(`should copy public html and replace the assetPrefix variable in rspack`, async () => {
+    await testPublicHtml('rspack');
+  });
+});
