+++ tests/integration/copy-assets/tests/tsconfig.json
